Portfolio_backend-work-controllers update
Added a check that the userId provided in the request in the auth tab is the same as the admin userId on the creation, modification, and deletion routes.

+++ backend/controllers/work.js
@@ -2,15 +2,19 @@ const Work = require('../models/work');
 
 exports.createWork = (req, res, next) => {
     const workObject = (req.body);
-    const work = new Work({
-        ...workObject
-    });
+    if (req.auth.userId != process.env.USER_ID) {
+        res.status(403).json({ message: 'Not authorized' });
+    } else {
+        const work = new Work({
+            ...workObject,
+        });
 
-    work.save()
-        .then(() => {
-            res.status(201).json({ message: 'travail enregistré !', })
-        })
-        .catch(error => { res.status(400).json({ error }) })
+        work.save()
+            .then(() => {
+                res.status(201).json({ message: 'travail enregistré !', })
+            })
+            .catch(error => { res.status(400).json({ error }) });
+    }
 };
 
 exports.getOneWork = (req, res, next) => {
@@ -34,19 +38,28 @@ exports.getOneWork = (req, res, next) => {
 
 exports.modifyWork = (req, res, next) => {
     const workObject = (req.body);
-    Work.updateOne({ _id: req.params.id }, { ...workObject, _id: req.params.id })
-        .then(() => res.status(200).json({
-            status: 'Projet modifié !',
-        }))
-        .catch(error => res.status(401).json({ error }));
+    if (req.auth.userId != process.env.USER_ID) {
+        res.status(403).json({ message: 'Not authorized' });
+    } else {
+        Work.updateOne({ _id: req.params.id }, { ...workObject, _id: req.params.id })
+            .then(() => res.status(200).json({
+                status: 'Projet modifié !',
+            }))
+            .catch(error => res.status(401).json({ error }));
+    }
 };
 
 exports.deleteWork = (req, res, next) => {
-    Work.deleteOne({ _id: req.params.id })
-        .then(() => {
-            res.status(200).json({ message: 'Projet supprimé !' })
-        })
-        .catch(error => res.status(401).json({ error }));
+
+    if (req.auth.userId != process.env.USER_ID) {
+        res.status(403).json({ message: 'Not authorized' });
+    } else {
+        Work.deleteOne({ _id: req.params.id })
+            .then(() => {
+                res.status(200).json({ message: 'Projet supprimé !' })
+            })
+            .catch(error => res.status(401).json({ error }));
+    }
 };
 
 exports.getAllWorks = (req, res, next) => {
